rename script files, remove print statements
