+++ app/db/script_create_tables.py
