+++ app/db/script_drop_data.py
