+++ app/db/script_reset_db.py
@@ -1,5 +1,5 @@
-import app.db.drop_data as drop
-import app.db.create_tables as create
+import app.db.script_drop_data as drop
+import app.db.script_create_tables as create
 
 def run():
     print("Dropping Tables")
@@ -7,9 +7,6 @@ def run():
 
     print("Creating Tables")
     create.create()
-
-    # print("Inserting Data")
-    # generate.make()
 
     print("\n===> Everything is completed\n")
